Version 1.2

+++ app/Http/Controllers/VagaController.php
@@ -4,15 +4,22 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Vaga;
-
+use Illuminate\Support\Facades\Auth;
 class VagaController extends Controller
 {
+
     public function index()
     {
-        $quantidadeDeItensAExibir = 5;
-        $listaDeVagas = Vaga::latest()->paginate($quantidadeDeItensAExibir);
-        
-        return view('vaga/index',compact('listaDeVagas'))->with('i', (request()->input('page', 1) - 1) * $quantidadeDeItensAExibir );
+        if(Auth::check()){
+
+            $quantidadeDeItensAExibir = 5;
+            $listaDeVagas = Vaga::latest()->paginate($quantidadeDeItensAExibir);
+            
+            return view('vaga/index',compact('listaDeVagas'))->with('i', (request()->input('page', 1) - 1) * $quantidadeDeItensAExibir );
+
+        }
+  
+        return redirect("login")->withSuccess('You are not allowed to access');
     }
 
     /**
@@ -22,7 +29,12 @@ class VagaController extends Controller
      */
     public function create()
     {
+        if(Auth::check()){
         return view('vaga/create');
+
+        }
+
+        return redirect("login")->withSuccess('You are not allowed to access');
     }
 
     /**
@@ -33,12 +45,23 @@ class VagaController extends Controller
      */
     public function store(Request $request)
     {
-        $storeData = $request->validate([
+        $request->validate([
             'titulo' => 'required|max:255',
             'descricao' => 'required|min:3',
-            'status' => 'required'
+            'status' => 'required',
+            'arquivo' => 'required',
         ]);
-        $vaga = Vaga::create($storeData);
+
+        $input = $request->all();
+  
+        if ($arquivo = $request->file('arquivo')) {
+            $destinationPath = 'arquivo/';
+            $profilearquivo = date('YmdHis') . "." . $arquivo->getClientOriginalExtension();
+            $arquivo->move($destinationPath, $profilearquivo);
+            $input['arquivo'] = "$profilearquivo";
+        }
+
+        $vaga = Vaga::create($input);
         return redirect('/vaga')
                 ->with('success', 'Vaga salvo com sucesso.');
     }
@@ -51,7 +74,11 @@ class VagaController extends Controller
      */
     public function show(Vaga $vaga)
     {
+        if(Auth::check()){
         return view('vaga/show', compact('vaga'));
+        }
+
+        return redirect("login")->withSuccess('You are not allowed to access');
     }
 
     /**
@@ -62,7 +89,10 @@ class VagaController extends Controller
      */
     public function edit(Vaga $vaga)
     {
+        if(Auth::check()){
         return view('vaga.edit',compact('vaga'));
+        }
+        return redirect("login")->withSuccess('You are not allowed to access');
     }
 
     /**
@@ -74,28 +104,28 @@ class VagaController extends Controller
      */
     public function update(Request $request, Vaga $vaga)
     {
-        {
-            $request->validate([
-                'name' => 'required',
-                'detail' => 'required'
-            ]);
-      
-            $input = $request->all();
-      
-            if ($image = $request->file('image')) {
-                $destinationPath = 'images/';
-                $profileImage = date('YmdHis') . "." . $image->getClientOriginalExtension();
-                $image->move($destinationPath, $profileImage);
-                $input['image'] = "$profileImage";
-            }else{
-                unset($input['image']);
-            }
-              
-            $product->update($input);
-        
-            return redirect()->route('products.index')
-                            ->with('success','Product updated successfully');
+        $request->validate(
+            [
+                'titulo' => 'required|max:255',
+                'descricao' => 'required|min:3',
+                'status' => 'required'
+            ]
+        );
+
+        $storeData =  $request->all();
+  
+        if ($arquivo = $request->file('arquivo')) {
+            $destinationPath = 'arquivo/';
+            $profilearquivo = date('YmdHis') . "." . $arquivo->getClientOriginalExtension();
+            $arquivo->move($destinationPath, $profilearquivo);
+            $storeData['arquivo'] = "$profilearquivo";
+        }else{
+            unset($storeData['arquivo']);
         }
+
+        $vaga->update($storeData); //Método Herdado
+
+        return redirect('/vaga')->with('success','Vaga Editada com Sucesso');
     }
 
     /**
@@ -106,7 +136,10 @@ class VagaController extends Controller
      */
     public function destroy(Vaga $vaga)
     {
+        if(Auth::check()){
         $vaga->delete();
         return redirect('/vaga')->with('success','Vaga Apagada com Sucesso');
+        }
+        return redirect("login")->withSuccess('You are not allowed to access');
     }
 }
